Made home view more atomic in its design and removed confusing comments
